Move MyPreferenceFragment out of Settings activity class
Commit:
https://github.com/duki994/Synapse/commit/464d92fe791fbccaa26740569fd8937e38a616ec fixed issue https://github.com/AndreiLux/Synapse/issues/52

That commit also made new bug (or rather exposed what was happening in https://github.com/AndreiLux/Synapse/issues/52).
On each rotation, Settings was unable to instantiate MyPreferenceFragment. Each time rotation changed while user was in settings menu, Synapse would crash due to InstantiationException

Fix this by moving MyPreferenceFragment out of Settings and making it outer standalone class.
