+++ src/main/java/com/af/synapse/Settings.java
@@ -7,13 +7,9 @@ import android.graphics.PorterDuff;
 import android.graphics.drawable.BitmapDrawable;
 import android.os.Bundle;
 import android.preference.PreferenceActivity;
-import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
-import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
 import android.view.Window;
 
 import com.af.synapse.utils.Utils;
@@ -128,23 +124,5 @@ public class Settings extends PreferenceActivity {
         }
 
         return super.onOptionsItemSelected(item);
-    }
-
-    public class MyPreferenceFragment extends PreferenceFragment {
-        @Override
-        public void onCreate(final Bundle savedInstanceState)
-        {
-            super.onCreate(savedInstanceState);
-            PreferenceManager.setDefaultValues(Settings.this, R.xml.preference_main, true);
-            addPreferencesFromResource(R.xml.preference_main);
-        }
-
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup root, Bundle savedInstanceState){
-            View v = super.onCreateView(inflater, root, savedInstanceState);
-            if (v != null)
-                v.setPadding(0, MainActivity.topPadding, 0, 0);
-            return v;
-        }
     }
 }
